+++ tests/store/test_datasource.py
@@ -1,4 +1,3 @@
-import datetime
 import uuid
 
 import numpy as np
@@ -10,7 +9,8 @@ from openghg.objectstore import get_bucket, exists
 from openghg.standardise.surface import parse_crds
 from openghg.objectstore import Datasource
 from openghg.types import ObjectStoreError, ZarrStoreError
-from openghg.util import create_daterange_str, timestamp_tzaware
+from openghg.types._errors import DataOverlapError
+from openghg.util import timestamp_tzaware
 
 mocked_uuid = "00000000-0000-0000-00000-000000000000"
 mocked_uuid2 = "10000000-0000-0000-00000-000000000001"
@@ -27,10 +27,10 @@ def create_attributes():
 
 # Disable this for long strings below - Line break occurred before a binary operator (W503)
 # flake8: noqa: W503
-def create_three_datasets(a, b, c):
-    values_a = np.arange(0, len(a), 1)
-    values_b = np.arange(0, len(b), 1)
-    values_c = np.arange(0, len(c), 1)
+def create_three_datasets(a, b, c, constant=False):
+    values_a = np.zeros(len(a)) if constant else np.arange(0, len(a), 1)
+    values_b = np.full(len(b), 1) if constant else np.arange(0, len(b), 1)
+    values_c = np.full(len(c), 2) if constant else np.arange(0, len(c), 1)
 
     attributes = create_attributes()
 
@@ -107,7 +107,7 @@ def datasets_with_overlap():
     time_b = pd.date_range("2012-01-29T00:00:00", "2012-04-30T00:00:00", freq="1d")
     time_c = pd.date_range("2012-04-16T00:00:00", "2012-09-30T00:00:00", freq="1d")
 
-    return create_three_datasets(time_a, time_b, time_c)
+    return create_three_datasets(time_a, time_b, time_c, constant=True)
 
 
 def test_add_data(data, datasource):
@@ -445,19 +445,8 @@ def test_add_data_with_gaps_check_stored_dataset(datasets_with_gaps, datasource)
 
 
 def test_add_data_with_overlap_check_stored_dataset(datasource, datasets_with_overlap):
-    time_a = pd.date_range("2012-01-01T00:00:00", "2012-01-31T00:00:00", freq="1d")
-    time_b = pd.date_range("2012-01-29T00:00:00", "2012-04-30T00:00:00", freq="1d")
-    time_c = pd.date_range("2012-04-16T00:00:00", "2012-09-30T00:00:00", freq="1d")
-
-    values_a = np.zeros(len(time_a))
-    values_b = np.full(len(time_b), 1)
-    values_c = np.full(len(time_c), 2)
-
-    attributes = create_attributes()
-
-    data_a = xr.Dataset({"mf": ("time", values_a)}, coords={"time": time_a}, attrs=attributes)
-    data_b = xr.Dataset({"mf": ("time", values_b)}, coords={"time": time_b}, attrs=attributes)
-    data_c = xr.Dataset({"mf": ("time", values_c)}, coords={"time": time_c}, attrs=attributes)
+    """Check that we can add data with overlaps."""
+    data_a, data_b, data_c = datasets_with_overlap
 
     attributes = create_attributes()
 
@@ -474,28 +463,42 @@ def test_add_data_with_overlap_check_stored_dataset(datasource, datasets_with_ov
         assert ds.equals(combined)
 
 
-def test_add_data_out_of_order(datasource, datasets_with_gaps):
-    data_a, data_b, data_c = datasets_with_gaps
+def test_error_if_overlap_and_not_combine(datasource, datasets_with_overlap):
+    """Check that we can get an error if we add overlapping data with if_exists == "auto"."""
+    data_a, data_b, _ = datasets_with_overlap
+
     attributes = create_attributes()
 
     d = datasource
 
+    d.add_data(metadata=attributes, data=data_a, data_type="surface", new_version=False)
+
+    with pytest.raises(DataOverlapError):
+        d.add_data(metadata=attributes, data=data_b, data_type="surface", new_version=False, if_exists="auto")
+
+
+# TODO: fix attribute handling so this doesn't happen!
+def test_attributes_overwritten_on_combine(datasource, datasets_with_overlap):
+    """Test that shows attributes are overwritten when writing to zarr.
+
+    We probably want to change this, but we'll need to add code to do it.
+
+    NOTE: the specific choice of datasets and the use of `if_exists == "combine"` doesn't
+    matter for this test.
+    """
+    data_a, data_b, _ = datasets_with_overlap
+
+    attributes = create_attributes()
+
+    d = datasource
+
+    d.add_data(metadata=attributes, data=data_a, data_type="surface", new_version=False)
+
+    data_b.attrs = {}
+
     d.add_data(metadata=attributes, data=data_b, data_type="surface", new_version=False, if_exists="combine")
-    d.add_data(metadata=attributes, data=data_a, data_type="surface", new_version=False, if_exists="combine")
-    d.add_data(metadata=attributes, data=data_c, data_type="surface", new_version=False, if_exists="combine")
 
-    assert d.data_keys() == [
-        "2012-01-01-00:00:00+00:00_2012-01-31-00:00:59+00:00",
-        "2012-04-01-00:00:00+00:00_2012-04-30-00:00:59+00:00",
-        "2012-09-01-00:00:00+00:00_2012-09-30-00:00:59+00:00",
-    ]
-
-    expected = xr.concat([data_a, data_b, data_c], dim="time").drop_duplicates("time", keep="last").sortby("time")
-
-    ds = d.get_data(version="v1").compute()
-
-    assert ds.time.size == expected.time.size
-    assert ds.equals(expected)
+    assert d.get_data().attrs == {}
 
 
 def test_add_data_out_of_order_no_combine(datasource, datasets_with_gaps):